CM-29960 - Migrate all CLI commands to new project structure
