+++ tests/cli/commands/configure/test_configure_command.py
