+++ tests/cli/files_collector/iac/test_tf_content_generator.py
